+++ plone/app/content/browser/constraintypes.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
+from plone.autoform.form import AutoExtensibleForm
 from Products.CMFPlone import PloneMessageFactory as PC_
 from Products.CMFPlone.interfaces import ISelectableConstrainTypes
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
-from plone.autoform.form import AutoExtensibleForm
 from z3c.form import button
 from z3c.form import form
 from z3c.form.browser.checkbox import CheckBoxFieldWidget
-from zope.interface import Interface
 from zope.interface import implementer
+from zope.interface import Interface
 from zope.interface import invariant
 from zope.interface.exceptions import Invalid
 from zope.schema import Choice
@@ -15,6 +15,7 @@ from zope.schema import List
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
+
 
 # XXX
 # acquire locallyAllowedTypes from parent (default)